Adopt common naming scheme
