+++ src/opencv_cam_main.cpp
@@ -1,7 +1,6 @@
-#include "opencv_cam/opencv_cam_component.hpp"
-#include "opencv_cam/subscriber_component.hpp"
+#include "opencv_cam/opencv_cam_node.hpp"
 
-// Manually compose OpencvCamNode and ImageSubscriberNode with use_intra_process_comms=true
+// Launch OpencvCamNode with use_intra_process_comms=true
 
 int main(int argc, char **argv)
 {
@@ -17,12 +16,8 @@ int main(int argc, char **argv)
   // Create and add camera node
   rclcpp::NodeOptions options{};
   options.use_intra_process_comms(true);
-  auto camera_node = std::make_shared<opencv_cam::OpencvCamNode>(options);
-  executor.add_node(camera_node);
-
-  // Create and add subscriber node
-  auto subscriber_node = std::make_shared<opencv_cam::ImageSubscriberNode>(options);
-  executor.add_node(subscriber_node);
+  auto node = std::make_shared<opencv_cam::OpencvCamNode>(options);
+  executor.add_node(node);
 
   // Spin until rclcpp::ok() returns false
   executor.spin();
